ai/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9

# Machine Learning - PyTorch
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# Satellite calculations
sgp4==2.23
pandas==2.1.3

# Utilities
python-dateutil==2.8.2